+++ src/methods/get_updates.rs
@@ -13,7 +13,7 @@ pub(crate) struct GetUpdates<'a, C> {
     #[serde(skip_serializing_if = "Option::is_none")]
     limit: Option<u8>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    timeout: Option<u32>,
+    timeout: Option<u64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     allowed_updates: Option<&'a [types::Updates]>,
 }
@@ -24,7 +24,7 @@ impl<'a, C> GetUpdates<'a, C> {
         token: Token,
         offset: Option<u32>,
         limit: Option<u8>,
-        timeout: Option<u32>,
+        timeout: Option<u64>,
         allowed_updates: Option<&'a [types::Updates]>,
     ) -> Self {
         Self {
